chore: stop plan on delete

+++ arex-web-api/src/main/java/com/arextest/web/api/service/controller/ReportQueryController.java
@@ -7,6 +7,7 @@ import javax.validation.Valid;
 import com.arextest.web.common.HttpUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -276,7 +277,12 @@ public class ReportQueryController {
     public Response deleteReport(@PathVariable String planId) {
         try {
             SuccessResponseType response = new SuccessResponseType();
-            HttpUtils.get(stopPlanUrl + "/?planId=" + planId, Response.class);
+            ResponseEntity<String> stopRes = HttpUtils.get(stopPlanUrl + "/?planId=" + planId, String.class);
+            if (!stopRes.hasBody() || StringUtils.isEmpty(stopRes.getBody())) {
+                LOGGER.error("stop plan error, planId:{}", planId);
+                return ResponseUtils.errorResponse("stop plan error", ResponseCode.REQUESTED_HANDLE_EXCEPTION);
+            }
+            LOGGER.info("stop plan success, {}", stopRes.getBody());
             response.setSuccess(reportService.deleteReport(planId));
             return ResponseUtils.successResponse(response);
         } catch (Throwable t) {
